Added ownship prediction
Ownship prediction is now available together with the cluster mode. It
uses the same timeframe of the traffic prediction.
It returns n points in the future where n is equal to the number of
steps indicated in the request. It is not a clusterization, it is a
single point given by a single instance of particles. But it can be
used as a region to check for conflicts